+++ app/src/main/java/com/example/togetherpet/fragment/WalkingPetResultFragment.kt
@@ -15,6 +15,7 @@ import androidx.lifecycle.lifecycleScope
 import androidx.lifecycle.repeatOnLifecycle
 import com.example.togetherpet.R
 import com.example.togetherpet.databinding.FragmentWalkingPetResultBinding
+import com.example.togetherpet.extensions.drawLine
 import com.kakao.vectormap.KakaoMap
 import com.kakao.vectormap.KakaoMapReadyCallback
 import com.kakao.vectormap.LatLng
@@ -78,22 +79,6 @@ class WalkingPetResultFragment : Fragment() {
         })
     }
 
-    fun drawLine(arrayList: ArrayList<LatLng>){
-        val layer = kakaoMap.routeLineManager?.layer
-        val lineStyle = RouteLineStyle.from(16f, Color.RED)
-        lineStyle.strokeColor = Color.BLACK
-        val stylesSet = RouteLineStylesSet.from(
-            RouteLineStyles.from(lineStyle)
-        )
-        val segment = RouteLineSegment.from(
-            arrayList
-        ).setStyles(stylesSet.getStyles(0))
-        val options = RouteLineOptions.from(segment)
-            .setStylesSet(stylesSet)
-
-        val routeLine = layer?.addRouteLine(options)
-    }
-
     fun displayStartPoint(arrayList: ArrayList<LatLng>) {
         createLabel(arrayList.first())
     }
@@ -117,7 +102,7 @@ class WalkingPetResultFragment : Fragment() {
                     Log.d("testt", "listener, Array : ${arrayLoc}")
                     if(arrayLoc.isNotEmpty()){
                         displayStartPoint(arrayLoc)
-                        drawLine(arrayLoc)
+                        kakaoMap.drawLine(arrayLoc)
                         displayEndPoint(arrayLoc)
                     }
                 }
